requirements.txt
# -*- coding: utf-8 -*-
# 工程预装模块
# 需要额外的python包，可直接在文件后面添加
# 请确保指定的包和版本号，可通过pip安装

Django==1.8.3
MySQL-python==1.2.3
MarkupSafe==0.23
Mako==1.0.1
httplib2==0.9.1
suds==0.4
celery==3.1.18
django-celery==3.1.16
requests==2.11.1
blueking.component.ieod==0.0.31
pycrypto==2.6.1
paramiko==1.16.0
ansible==1.9.4
rpyc==3.3.0
bkstorages==0.1.1
djangorestframework==3.6.2
django-filter>=1.0.2
django-cors-headers